Pickers: simplified structure with polymorphism instead of composition
