+++ src/MudBlazor/Components/DatePicker/MudDateRangePicker.razor
@@ -1,0 +1,16 @@
+﻿@namespace MudBlazor
+@inherits MudBaseDatePicker
+
+@Render()
+
+@code {
+    protected override RenderFragment InputContent()
+    {
+        return @<MudInputControl Label="@Label" HelperText="@HelperText" Variant="@InputVariant" Disabled="@Disabled" @onclick="() => { if (!Editable) ToggleState(); }">
+                   <InputContent>
+                       <MudRangeInput @attributes="UserAttributes" InputType="@InputType.Text" Class="@PickerInputClass" Style="@Style" Variant="@InputVariant" ReadOnly="@(!Editable)" T="string"
+                                       @bind-Text="@(Text)" Disabled=@Disabled Adornment="@Adornment" AdornmentIcon="@InputIcon" IconSize="@IconSize" OnAdornmentClick="ToggleState"/>
+                   </InputContent>
+               </MudInputControl>;
+    }
+}
