+++ src/MudBlazor/Utilities/InternalComponents/Node.razor
@@ -1,0 +1,8 @@
+﻿@using Microsoft.AspNetCore.Components
+@namespace MudBlazor.Internal
+
+@ChildContent
+
+@code {
+    [Parameter] public RenderFragment ChildContent { get; set; }
+}
